+++ src/nemdb/dnsp/essential_energy.py
@@ -49,7 +49,7 @@ def get_url(year: int):
         return r.content
 
 
-@pa.check_output(LoadSchema)
+@pa.check_output(LoadSchema)  # type: ignore[arg-type]
 def _read_all_zss(file):
     """
     Read a zip file containing csvs of load data for each zone substation (ZSS)
@@ -79,6 +79,5 @@ def _read_all_zss(file):
 
 
 if __name__ == "__main__":
-    file = "/home/simba/Downloads/EE-Zone-Substation-Load-Data-2023-24.zip"
-    df = read_all_zss(file)
+    df = read_all_zss(2024)
     print(df)