+++ src/nemdb/dnsp/tasnetworks.py
@@ -19,7 +19,7 @@ def get_url(year: int):
     }.get(year)
 
 
-@pa.check_output(LoadSchema)
+@pa.check_output(LoadSchema)  # type: ignore[arg-type]
 def _read_all_zss(file):
     df = (
         pl.from_pandas(
@@ -49,6 +49,5 @@ def _read_all_zss(file):
 
 
 if __name__ == "__main__":
-    path = "/home/simba/Downloads/Tasnetworks-Zone-Substation-Load-Data-2023-24.csv"
-    df = read_all_zss(path)
+    df = read_all_zss(2024)
     print(df)